# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hardware AI Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  draft: "draft",
  in_progress: "in_progress",
  ready_to_order: "ready_to_order",
  ordered: "ordered",
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  /** @nullable */
  description?: string | null;
}

/**
 * @nullable
 */
export type UpdateProjectBodyStatus =
  | (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus]
  | null;

export const UpdateProjectBodyStatus = {
  draft: "draft",
  in_progress: "in_progress",
  ready_to_order: "ready_to_order",
  ordered: "ordered",
} as const;

export interface UpdateProjectBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  status?: UpdateProjectBodyStatus;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface Message {
  id: number;
  projectId: number;
  role: MessageRole;
  content: string;
  /**
   * Base64-encoded image data (without data URL prefix)
   * @nullable
   */
  imageData?: string | null;
  /**
   * MIME type for imageData, e.g. image/png
   * @nullable
   */
  imageMediaType?: string | null;
  createdAt: string;
}

export interface SendMessageBody {
  content: string;
  /**
   * Optional base64-encoded image (no data URL prefix)
   * @nullable
   */
  imageData?: string | null;
  /**
   * MIME type of imageData, e.g. image/png
   * @nullable
   */
  imageMediaType?: string | null;
}

export interface PartSpec {
  id: number;
  projectId: number;
  /** e.g. bracket, plate, enclosure, channel, angle */
  partType: string;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  thickness?: number | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  depth?: number | null;
  /** @nullable */
  bendRadius?: number | null;
  /**
   * JSON array of bend angles in degrees
   * @nullable
   */
  bendAngles?: string | null;
  /**
   * JSON describing hole locations and diameters
   * @nullable
   */
  holePattern?: string | null;
  /** @nullable */
  powderCoat?: boolean | null;
  /** @nullable */
  powderCoatColor?: string | null;
  /** @nullable */
  notes?: string | null;
  /**
   * SVG string for 2D flat pattern preview
   * @nullable
   */
  svgPreview?: string | null;
  /** @nullable */
  sendCutSendUrl?: string | null;
  /**
   * Canonical parametric DSL (JSON-encoded) describing the part
   * @nullable
   */
  dslJson?: string | null;
  /**
   * Generated feature graph (JSON-encoded) used to drive preview and DXF
   * @nullable
   */
  featureGraphJson?: string | null;
  /** @nullable */
  currentRevisionId?: number | null;
  totalRevisions: number;
  updatedAt: string;
}

export type RuleResultStatus =
  (typeof RuleResultStatus)[keyof typeof RuleResultStatus];

export const RuleResultStatus = {
  pass: "pass",
  warn: "warn",
  fail: "fail",
  na: "na",
} as const;

/**
 * @nullable
 */
export type RuleResultSuggestion = { [key: string]: unknown } | null;

export interface RuleResult {
  id: string;
  label: string;
  status: RuleResultStatus;
  message: string;
  /** @nullable */
  suggestion?: RuleResultSuggestion;
}

export type ValidationSnappedSpec = { [key: string]: unknown };

export interface Validation {
  rules: RuleResult[];
  hasFailures: boolean;
  snappedSpec: ValidationSnappedSpec;
}

export interface ChatResponse {
  message: Message;
  partSpec?: PartSpec;
  partUpdated: boolean;
  validation?: Validation;
}

export interface ApplySuggestionResponse {
  partSpec: PartSpec;
  validation: Validation;
}

export interface UpdatePartSpecBody {
  /** @nullable */
  partType?: string | null;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  thickness?: number | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  depth?: number | null;
  /** @nullable */
  bendRadius?: number | null;
  /** @nullable */
  bendAngles?: string | null;
  /** @nullable */
  holePattern?: string | null;
  /** @nullable */
  powderCoat?: boolean | null;
  /** @nullable */
  powderCoatColor?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface DxfExport {
  projectId: number;
  filename: string;
  /** DXF file content as a string */
  dxfContent: string;
  sendCutSendUploadUrl: string;
  instructions: string[];
  estimatedParts: number;
}

export type MaterialCategory =
  (typeof MaterialCategory)[keyof typeof MaterialCategory];

export const MaterialCategory = {
  steel: "steel",
  aluminum: "aluminum",
  stainless: "stainless",
  copper: "copper",
  brass: "brass",
  other: "other",
} as const;

export interface Material {
  id: number;
  name: string;
  category: MaterialCategory;
  sendCutSendName: string;
  canBend: boolean;
  canPowderCoat: boolean;
  minThickness: number;
  maxThickness: number;
  maxSheetWidth: number;
  maxSheetHeight: number;
  /** @nullable */
  description?: string | null;
}

export interface Thickness {
  id: number;
  materialId: number;
  /** @nullable */
  gauge?: string | null;
  inches: number;
  mm: number;
}

export type CapabilitiesCutting = {
  method: string;
  maxWidth: number;
  maxHeight: number;
  minFeatureSize: number;
};

export type CapabilitiesBending = {
  maxLength: number;
  minFlange: number;
  tolerances: string;
};

export interface Capabilities {
  cutting: CapabilitiesCutting;
  bending: CapabilitiesBending;
  finishing: string[];
  fileFormats: string[];
  uploadUrl: string;
  turnaround: string;
}

export interface PartRevision {
  id: number;
  projectId: number;
  revisionNumber: number;
  dslJson: string;
  /** @nullable */
  rationale?: string | null;
  createdAt: string;
}

export interface RevisionPreview {
  revision: PartRevision;
  partSpec: PartSpec;
  isCurrent: boolean;
}

export interface RevisionChangeResponse {
  partSpec: PartSpec;
  /** @nullable */
  currentRevisionId?: number | null;
  canUndo: boolean;
  canRedo: boolean;
}

export interface ProjectSummary {
  id: number;
  name: string;
  status: string;
  /** @nullable */
  partType?: string | null;
  /** @nullable */
  material?: string | null;
  updatedAt: string;
}